Improve parser regex, use multiprocessing to divide and conquer the dump files
Regex for extracting sections did not account for images placed in a
different order after the title line.

Use multiprocessing and partitioning of the dump file to speed up the
process. File is partitioned by as many processors as the user has.
Partitioning only occurs on blank lines, so as to not interrupt the
pattern the regex is looking for. JSON files are generated for each
partition and then combined back together at the end. This took
the processing time for the French dump from ~20 minutes to 5
on a quad-core Intel i5-3570 processor at 3.4 GHz. This mode
will be the default unless user specifies that they want a single
process using -p/--single-process.